+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Note: version releases in the 0.x.y range may introduce breaking changes.
 
+## [unreleased]
+ ### Added 
+ ### Fixed 
+
 ## [2.16.0]
  ### Added 
  ### Fixed 
@@ -434,3 +438,4 @@ Note: version releases in the 0.x.y range may introduce breaking changes.
 [2.14.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.13.0...v2.14.0
 [2.15.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.14.0...v2.15.0
 [2.16.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.15.0...v2.16.0
+[unreleased]: https://github.com/ehrbase/openEHR_SDK/compare/v2.16.0...HEAD